lib: untangle export paths for MID and flight clients
