+++ packages/lib/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@desci-labs/desci-codex-lib",
-  "version": "2.0.0-rc5",
+  "version": "2.0.0-rc6",
   "description": "Codex interaction primitives",
   "license": "MIT",
   "author": "Edvard Hübinette",
