--- packages/lib/src/c1/clients.ts
@@ -1,25 +1,0 @@
-import {
-  createFlightSqlClient,
-  FlightSqlClient,
-} from "@ceramic-sdk/flight-sql-client";
-import { ModelInstanceClient } from "@ceramic-sdk/model-instance-client";
-import type { StreamClientParams } from "@ceramic-sdk/stream-client";
-
-export const DEFAULT_LOCAL_FLIGHT = "http://localhost:5102";
-
-export const newFlightSqlClient = (flightUrl: string) => {
-  const url = new URL(flightUrl);
-
-  return createFlightSqlClient({
-    host: url.hostname,
-    port: parseInt(url.port),
-    tls: url.protocol === "https:",
-    headers: [],
-  });
-};
-
-export { FlightSqlClient };
-
-export const newStreamClient = (config: StreamClientParams) => {
-  return new ModelInstanceClient(config);
-};
