+++ packages/lib/src/c1/index.ts
@@ -2,5 +2,6 @@ export * from "./sql.js";
 export * from "./resolve.js";
 export * from "./explore.js";
 export * from "./mutate.js";
-export * from "./clients.js";
+export * from "./flightclient.js";
+export * from "./streamclient.js";
 export * from "./queries.js";
